MAINT: Refactor memos array
